LOOKS: Make each status icon unique to avoid confusion
Intention is to prevent student's thinking that they're done with
discuss and demo being ticks. If they are not ticks, that will give
them insentive to actually go to their tutor and talk to them which
is exactly what we want them to do.

If they are ticks, students might think that they are done when
they really are not... leading to student's having overdue tasks
as they are not checked off as a green tick.

Also changes the redo to a portfolio icon, meaning put it in your
portfolio and never show me this task again.

Also changes not started to stop, instead of an 'x', meaning that
your work on this is paused/stale and you should get on with it.